Create initial Sphinx documentation skeleton
This is close to the base sphinx-quickstart setup, except the conf.py
file is tidied up with automatic versioning and a base set of extensions
for parsing docstrings and readthedocs formatting.

There is also a GitHub Action to ensure that the documentation is tested
on each pull request, so that improperly formatted dosctrings can't
sneak through.