Added tests for both TfTenso128 and TfTensor64
This is done modifying the prefious tests to rely on the class
attributes Datatype and dtype that specifie the creation dtype and the
expected dtype of the wrapped tensor. In this way the tests for
TfTensor64 can be inherited from TfTensor128.

These tests required some minor modifications. The most notable one is
that some have to rely in comparisons between tf.tensors directly
instead of using the array method to compare numpy arrays. This si
because the array method always returns a complex128 dtype, regardless
of the underlying tf.tensor.